+++ sentry-samples/sentry-samples-spring/src/main/java/io/sentry/samples/spring/AppInitializer.java
@@ -2,7 +2,6 @@ package io.sentry.samples.spring;
 
 import io.sentry.spring.tracing.SentryTracingFilter;
 import javax.servlet.Filter;
-
 import org.springframework.web.filter.DelegatingFilterProxy;
 import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;
 
@@ -25,8 +24,12 @@ public class AppInitializer extends AbstractAnnotationConfigDispatcherServletIni
 
   @Override
   protected Filter[] getServletFilters() {
+    // creates Sentry transactions around incoming HTTP requests
+    SentryTracingFilter sentryTracingFilter = new SentryTracingFilter();
+
+    // filter required by Spring Security
     DelegatingFilterProxy springSecurityFilterChain = new DelegatingFilterProxy();
     springSecurityFilterChain.setTargetBeanName("springSecurityFilterChain");
-    return new Filter[] {new SentryTracingFilter(), springSecurityFilterChain};
+    return new Filter[] {sentryTracingFilter, springSecurityFilterChain};
   }
 }
