+++ sentry-spring/api/sentry-spring.api
@@ -86,6 +86,16 @@ public class io/sentry/spring/tracing/SentrySpanClientHttpRequestInterceptor : o
 	public fun intercept (Lorg/springframework/http/HttpRequest;[BLorg/springframework/http/client/ClientHttpRequestExecution;)Lorg/springframework/http/client/ClientHttpResponse;
 }
 
+public class io/sentry/spring/tracing/SentryTracingConfiguration {
+	public fun <init> ()V
+	public fun sentrySpanAdvice (Lio/sentry/IHub;)Lorg/aopalliance/aop/Advice;
+	public fun sentrySpanAdvisor (Lio/sentry/IHub;Lorg/springframework/aop/Pointcut;)Lorg/springframework/aop/Advisor;
+	public fun sentrySpanPointcut ()Lorg/springframework/aop/Pointcut;
+	public fun sentryTransactionAdvice (Lio/sentry/IHub;)Lorg/aopalliance/aop/Advice;
+	public fun sentryTransactionAdvisor (Lio/sentry/IHub;Lorg/springframework/aop/Pointcut;)Lorg/springframework/aop/Advisor;
+	public fun sentryTransactionPointcut ()Lorg/springframework/aop/Pointcut;
+}
+
 public class io/sentry/spring/tracing/SentryTracingFilter : org/springframework/web/filter/OncePerRequestFilter {
 	public fun <init> ()V
 	public fun <init> (Lio/sentry/IHub;Lio/sentry/spring/SentryRequestResolver;)V
