+++ sentry-spring/src/test/kotlin/io/sentry/spring/tracing/SentrySpanAdviceTest.kt
@@ -15,12 +15,7 @@ import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
-import org.aopalliance.aop.Advice
 import org.junit.runner.RunWith
-import org.springframework.aop.Advisor
-import org.springframework.aop.Pointcut
-import org.springframework.aop.support.DefaultPointcutAdvisor
-import org.springframework.aop.support.annotation.AnnotationMatchingPointcut
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
